pom.xml toegevoegd. Projectbestanden verwijderd uit git. Out/ map verwijderd uit git Projectstructuur naar algemene projectstructuur omgeschreven.
