+++ .gitignore
@@ -24,7 +24,6 @@
 hs_err_pid*
 replay_pid*
 
-### Example user template template
 ### Example user template
 
 # IntelliJ project files
