--- src/AuthObserver.java
@@ -1,10 +1,0 @@
-public class AuthObserver implements ObserverListener {
-    @Override
-    public void update(String username, boolean isAuthenticated) {
-        if (isAuthenticated) {
-            System.out.println(username + " is successfully authenticated.");
-        } else {
-            System.out.println(username + " failed authentication.");
-        }
-    }
-}
