--- src/CSVReader.java
@@ -1,67 +1,0 @@
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
-import java.util.HashMap;
-import java.util.Map;
-
-public class CSVReader {
-    private String filePath;
-
-    public CSVReader(String filePath) {
-        this.filePath = filePath;
-    }
-
-    public Map<String, String[]> readAccounts() {
-        Map<String, String[]> accounts = new HashMap<>();
-        try (BufferedReader br = new BufferedReader(new FileReader(filePath))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-                String[] values = line.split(",");
-                if (values.length == 3) {
-                    accounts.put(values[0].trim(), new String[]{values[0].trim(), values[1].trim(), values[2].trim()});
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return accounts;
-    }
-
-
-
-    public void dataReader(){
-        String filePath = "Project-Halberd\\src\\data.csv";
-        String line;
-        String csvSplitBy = ",";
-
-        try (BufferedReader br = new BufferedReader(new FileReader(filePath))) {
-            while ((line = br.readLine()) != null) {
-                String[] values = line.split(csvSplitBy);
-                for (String value : values) {
-                    System.out.print(value + " ");
-                }
-                System.out.println();
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-    public String readAllContent() {
-        StringBuilder content = new StringBuilder();
-        try (BufferedReader br = new BufferedReader(new FileReader(filePath))) {
-            String line;
-            while ((line = br.readLine()) != null) {
-                String[] values = line.split(",");
-                if (values.length >= 3) {
-                    String username = values[0];
-                    String email = values[2];
-                    content.append(username).append("\n");
-                    content.append(email).append("\n");
-                }
-            }
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        return content.toString();
-    }
-}
