--- src/ObserverListener.java
@@ -1,4 +1,0 @@
-public interface ObserverListener {
-    void update(String username, boolean isAuthenticated);
-}
-
