--- src/SoundPlayer.java
@@ -1,17 +1,0 @@
-import javax.sound.sampled.*;
-import java.io.File;
-import java.io.IOException;
-
-public class SoundPlayer {
-    public static void playSound(String filePath) {
-        try {
-            File soundFile = new File(filePath);
-            AudioInputStream audioInputStream = AudioSystem.getAudioInputStream(soundFile);
-            Clip clip = AudioSystem.getClip();
-            clip.open(audioInputStream);
-            clip.start();
-        } catch (UnsupportedAudioFileException | IOException | LineUnavailableException e) {
-            e.printStackTrace();
-        }
-    }
-}
