+++ src/main/java/ObserverListener.java
@@ -1,0 +1,4 @@
+public interface ObserverListener {
+    void update(String username, boolean isAuthenticated);
+}
+
